+++ app/mypage/profile/page.tsx
@@ -1,69 +1,84 @@
 "use client"
 
-import { useState, useEffect } from "react"
+import { useEffect, useState } from "react"
+import { useRouter } from "next/navigation"
+import Link from "next/link"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Textarea } from "@/components/ui/textarea"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Label } from "@/components/ui/label"
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import { Badge } from "@/components/ui/badge"
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import { Alert, AlertDescription } from "@/components/ui/alert"
-import { X, Plus, Upload, Save, ArrowLeft, CheckCircle, User } from "lucide-react"
+import {
+  X,
+  Plus,
+  Upload,
+  Save,
+  ArrowRight,
+  CheckCircle,
+  Loader2,
+  User,
+  Heart,
+  Code,
+  MapPin,
+  Sparkles,
+  SkipBackIcon as Skip,
+} from "lucide-react"
 import Header from "@/components/header"
 import Footer from "@/components/footer"
+import { Profile, Skills, useAuth, UserSkills } from "@/contexts/auth-context"
 import ProtectedRoute from "@/components/protected-route"
-import { Profile, Skills, UserSkills, useAuth } from "@/contexts/auth-context"
-import Link from "next/link"
-import { useRouter } from "next/navigation"
-
-
-const defaultImage = "/placeholder.svg"
 
 function ProfileEditContent() {
+   const { viewProfile, saveProfile, user, 
+     updateUser, viewUserSkills, saveUserSkills,  getSkills } = useAuth()
 
-  const { viewProfile, saveProfile, isAuthenticated, user, 
-    updateUser, viewUserSkills, saveUserSkills,  getSkills } = useAuth()
-  //const [newInterest, setNewInterest] = useState("")
-  const [isLoading, setIsLoading] = useState(false)
-  //const [success, setSuccess] = useState(false)
+   const [isLoading, setIsLoading] = useState(false)
+ 
+   //선택된 스킬 데이터 저장
+   const [selectSkill, setSelectSkill] = useState<Skills[]>([]); // 빈 Skills 배열로 초기화
+ 
+   //모든 스킬 데이터를 가지고 있음
+   const [arraySkills, setArraySkills] = useState<Skills[]>([])
+ 
+   const [profile, setProfile] = useState<Profile>({
+     fullName: "",
+     bio: "",
+     profileImageUrl: "/placeholder.svg",
+     education: "",
+     experience: "",
+     portfolioUrl: "",
+   })
+ 
+   const [isSelectOpen, setIsSelectOpen] = useState(false); // Select 컴포넌트의 열림 상태를 관리하는 새 상태
+ 
+ 
+   const router = useRouter()
+   
+   useEffect(() => {
 
-  //선택된 스킬 데이터 저장
-  const [selectSkill, setSelectSkill] = useState<Skills[]>([]); // 빈 Skills 배열로 초기화
+     setIsLoading(true)     
 
-  //모든 스킬 데이터를 가지고 있음
-  const [arraySkills, setArraySkills] = useState<Skills[]>([])
+     if (user) { // 인증되었고 사용자가 존재할 때만 데이터 가져오기
 
-  const [profile, setProfile] = useState<Profile>({
-    fullName: "",
-    bio: "",
-    profileImageUrl: "/placeholder.svg",
-    education: "",
-    experience: "",
-    portfolioUrl: "",
-  })
+       fetchUserData(); 
 
-  const [isSelectOpen, setIsSelectOpen] = useState(false); // Select 컴포넌트의 열림 상태를 관리하는 새 상태
-
-
-  const router = useRouter()
+     } else {
   
-  useEffect(() => {
-    setIsLoading(true)
-    
-
-    if (isAuthenticated && user) { // 인증되었고 사용자가 존재할 때만 데이터 가져오기
-      fetchUserData();     
-    }else
       router.push("/")
 
-    setIsLoading(false)
-    
-    
-  }, [isAuthenticated, user]); // isAuthenticated와 user에 의존
+     }
+       
+     setIsLoading(false)
+     
+     
+   }, [user]); // 와 user에 의존
 
-  //프로필 및 스킬 정보를 가져옴
+
+   //프로필 및 스킬 정보를 가져옴
   const fetchUserData = async () => {
 
     try {
@@ -108,6 +123,7 @@ function ProfileEditContent() {
       console.error("프로필 또는 스킬 정보 로딩 중 오류:", error);
     }
   };
+
 
 
   // 프로필 저장 핸들러
@@ -173,237 +189,195 @@ function ProfileEditContent() {
     setSelectSkill((prev) => prev.filter((skill) => skill.id !== id));
   }
 
-  if (!user) return null
-
-    // const addInterest = (interest: string) => {
-  //   if (interest && !profile.interests.includes(interest)) {
-  //     setProfile({
-  //       ...profile,
-  //       interests: [...profile.interests, interest],
-  //     })
-  //   }
-  //   setNewInterest("")
-  // }
-
-  // const removeInterest = (interest: string) => {
-  //   setProfile({
-  //     ...profile,
-  //     interests: profile.interests.filter((i) => i !== interest),
-  //   })
-  // }
+  const handleSkip = () => {
+    router.push("/")
+  }
+  
+  if (!user) {
+    
+    return (
+      <div className="min-h-screen bg-gray-50 flex items-center justify-center">
+        <Card className="w-96">
+          <CardContent className="p-8 text-center">
+            <Loader2 className="w-8 h-8 animate-spin mx-auto mb-4 text-blue-600" />
+            <h3 className="text-lg font-medium text-gray-900 mb-2">로딩 중...</h3>
+            <p className="text-gray-600">사용자 정보를 확인하고 있습니다.</p>
+          </CardContent>
+        </Card>
+      </div>
+    )
+  }
+  if (isLoading) {
+    return (
+      <div className="min-h-screen bg-gray-50 flex items-center justify-center">
+        <Card className="w-96">
+          <CardContent className="p-8 text-center">
+            <Loader2 className="w-8 h-8 animate-spin mx-auto mb-4 text-blue-600" />
+            <h3 className="text-lg font-medium text-gray-900 mb-2">로딩 중...</h3>
+            <p className="text-gray-600">프로필 정보를 불러오는 중입니다.</p>
+          </CardContent>
+        </Card>
+      </div>
+    )
+  }
 
   return (
     <div className="min-h-screen bg-gray-50">
       <Header />
 
-      <div className="container mx-auto px-4 py-8">        
+      <div className="container mx-auto px-4 py-8">
         {/* 헤더 */}
-        <div className="flex items-center justify-between mb-8">
-          <Link href="/mypage">
-            <Button variant="outline" size="sm">
-              <ArrowLeft className="w-4 h-4 mr-2" />
-              돌아가기
-            </Button>
-          </Link>
-
-          <div className="flex items-center gap-4">
-            <div>
-              <h1 className="text-3xl font-bold text-gray-900">프로필 수정</h1>
-              <p className="text-gray-600">개인 정보와 관심사를 업데이트하세요</p>
-            </div>
-          </div>
+        <div className="text-center mb-8">
+          <h1 className="text-3xl font-bold text-gray-900 mb-2">프로필 설정</h1>
+          <p className="text-gray-600">{user.username}님만의 프로필을 만들어 더 정확한 공모전 추천을 받아보세요</p>
         </div>
 
-        {/* 성공 메시지 */}
-        {/* {success && (
-          <Alert className="mb-6 border-green-200 bg-green-50">
-            <CheckCircle className="h-4 w-4 text-green-600" />
-            <AlertDescription className="text-green-800">프로필이 성공적으로 업데이트되었습니다!</AlertDescription>
-          </Alert>
-        )} */}
-
-        <div className="grid grid-cols-1 lg:grid-cols-3 gap-8">
-          {/* 프로필 사진 */}
-          <div className="lg:col-span-1">
-            <Card>
-              <CardHeader>
-                <CardTitle>프로필 사진</CardTitle>
-              </CardHeader>
-              <CardContent className="text-center space-y-4">
-                <Avatar className="w-32 h-32 mx-auto">
-                  <AvatarImage src={profile?.profileImageUrl || defaultImage} alt={profile?.fullName} />
-                  <AvatarFallback className="text-4xl">{ profile?.fullName?.[0] ?? "U" }</AvatarFallback>
-                </Avatar>
-                <Button variant="outline" className="w-full bg-transparent" disabled>
-                  <Upload className="w-4 h-4 mr-2" />
-                  사진 업로드 (준비중)
-                </Button>
-                <p className="text-sm text-gray-500">JPG, PNG 파일만 업로드 가능합니다. (최대 5MB)</p>
-              </CardContent>
-            </Card>
-          </div>
-
-          {/* 프로필 정보 */}
-          <div className="lg:col-span-2 space-y-6">
-            {/* 기본 정보 */}
-            <Card>
-              <CardHeader>
-                <CardTitle>기본 정보</CardTitle>
-              </CardHeader>
-              <CardContent className="space-y-4">
-                <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                  <div className="space-y-2">
-                    <Label htmlFor="name">이름</Label>
-                    <Input
-                      id="name"
-                      value={profile?.fullName}
-                      onChange={(e) => setProfile({ ...profile, fullName: e.target.value })}
-                    />
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="email">이메일</Label>
-                    <Input
-                      id="email"
-                      type="email"
-                      value={user.email}
-                      // onChange={(e) => updateUser({ ...thisProfile, email: e.target.value })}
-                      disabled
-                      className="bg-gray-50"
-                    />
-                    <p className="text-xs text-gray-500">이메일은 변경할 수 없습니다.</p>
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="phone">전화번호</Label>
-                    <Input
-                      id="phone"
-                      value={user.phoneNumber}
-                      onChange={(e) => updateUser({ ...user, phoneNumber: e.target.value })}
-                      placeholder="010-0000-0000"
-                    />
-                  </div>
-                  {/* <div className="space-y-2">
-                    <Label htmlFor="location">지역</Label>
-                    <Select
-                      value={profile.location}
-                      onValueChange={(value) => updateUser({ ...profile, location: value })}
-                    >
-                      <SelectTrigger>
-                        <SelectValue placeholder="지역을 선택하세요" />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="서울">서울</SelectItem>
-                        <SelectItem value="부산">부산</SelectItem>
-                        <SelectItem value="대구">대구</SelectItem>
-                        <SelectItem value="인천">인천</SelectItem>
-                        <SelectItem value="광주">광주</SelectItem>
-                        <SelectItem value="대전">대전</SelectItem>
-                        <SelectItem value="울산">울산</SelectItem>
-                        <SelectItem value="세종">세종</SelectItem>
-                        <SelectItem value="경기">경기</SelectItem>
-                        <SelectItem value="강원">강원</SelectItem>
-                        <SelectItem value="충북">충북</SelectItem>
-                        <SelectItem value="충남">충남</SelectItem>
-                        <SelectItem value="전북">전북</SelectItem>
-                        <SelectItem value="전남">전남</SelectItem>
-                        <SelectItem value="경북">경북</SelectItem>
-                        <SelectItem value="경남">경남</SelectItem>
-                        <SelectItem value="제주">제주</SelectItem>
-                      </SelectContent>
-                    </Select>
-                  </div> */}
+        <div className="max-w-4xl mx-auto">
+          <Card>
+            <CardHeader>
+              <div className="flex justify-between items-center">
+                <div>
+                  <CardTitle className="text-2xl">종합 프로필 정보</CardTitle>
+                  <p className="text-sm text-gray-600">모든 프로필 정보를 한 곳에서 입력하고 관리하세요.</p>
                 </div>
+                <Button variant="ghost" onClick={handleSkip} className="shrink-0">
+                  나중에 작성하기
+                </Button>
+              </div>
+            </CardHeader>
+            <CardContent className="space-y-8 pt-6">
+              {/* 기본 정보 섹션 */}
+              <div className="space-y-6 p-6 border rounded-lg">
+                <h3 className="text-lg font-semibold flex items-center"><User className="w-5 h-5 mr-2" />기본 정보</h3>
+                {/* 프로필 사진 */}
+                <div className="text-center">
+                  <Avatar className="w-24 h-24 mx-auto mb-4">
+                    <AvatarImage src={"/placeholder.svg"} alt={user.username} />
+                    <AvatarFallback className="text-2xl">{user.username[0]}</AvatarFallback>
+                  </Avatar>
+                  <Button variant="outline" size="sm" disabled className="bg-transparent">
+                    <Upload className="w-4 h-4 mr-2" />
+                    사진 업로드 (준비중)
+                  </Button>
+                  <p className="text-xs text-gray-500 mt-2">JPG, PNG 파일만 업로드 가능합니다</p>
+                </div>
+
+                {/* 자기소개 */}
                 <div className="space-y-2">
                   <Label htmlFor="bio">자기소개</Label>
                   <Textarea
                     id="bio"
-                    value={profile?.bio}
-                     onChange={(e) => setProfile(prev => ({ ...prev, bio: e.target.value }))} // ✅ 항상 Profile 반환
+                    value={profile.bio}
+                    onChange={(e) => setProfile({ ...profile, bio: e.target.value })}
+                    placeholder="자신을 간단히 소개해주세요..."
                     rows={4}
-                    placeholder="자신을 소개해주세요..."
                   />
+                  <p className="text-xs text-gray-500">팀원들에게 보여질 소개글입니다</p>
                 </div>
-              </CardContent>
-            </Card>
 
-            {/* 관심 분야 */}
-            {/* <Card>
-              <CardHeader>
-                <CardTitle>관심 분야</CardTitle>
-              </CardHeader>
-              <CardContent className="space-y-4">
-                <div className="flex flex-wrap gap-2">
-                  {profile.interests.map((interest) => (
-                    <Badge key={interest} variant="secondary" className="flex items-center gap-1">
-                      {interest}
-                      <button
-                        onClick={() => removeInterest(interest)}
-                        className="ml-1 hover:bg-gray-300 rounded-full p-0.5"
-                      >
-                        <X className="w-3 h-3" />
-                      </button>
-                    </Badge>
-                  ))}
-                </div>
-                <div className="flex gap-2">
-                  <Select value={newInterest} onValueChange={setNewInterest}>
-                    <SelectTrigger className="flex-1">
-                      <SelectValue placeholder="관심 분야 선택" />
+                {/* 지역 */}
+                {/* <div className="space-y-2">
+                  <Label htmlFor="location" className="flex items-center">
+                    <MapPin className="w-4 h-4 mr-1" />
+                    활동 지역
+                  </Label>
+                  <Select
+                    value={profile.location}
+                    onValueChange={(value) => setProfile({ ...profile, location: value })}
+                  >
+                    <SelectTrigger>
+                      <SelectValue placeholder="지역을 선택하세요" />
                     </SelectTrigger>
                     <SelectContent>
-                      {availableInterests
-                        .filter((interest) => !profile.interests.includes(interest))
-                        .map((interest) => (
-                          <SelectItem key={interest} value={interest}>
-                            {interest}
-                          </SelectItem>
-                        ))}
+                      {regions.map((region) => (
+                        <SelectItem key={region} value={region}>
+                          {region}
+                        </SelectItem>
+                      ))}
                     </SelectContent>
                   </Select>
-                  <Button onClick={() => addInterest(newInterest)} disabled={!newInterest} size="sm">
-                    <Plus className="w-4 h-4" />
-                  </Button>
-                </div>
-              </CardContent>
-            </Card> */}
+                  <p className="text-xs text-gray-500">주로 활동하는 지역을 선택해주세요</p>
+                </div> */}
+              </div>
 
-            {/* 기술 스택 */}
-            <Card>
-              <CardHeader>
-                <CardTitle>기술 스택</CardTitle>
-              </CardHeader>
-              <CardContent className="space-y-4">
-                <div className="flex flex-wrap gap-2">
-                   {selectSkill?.map((skill) => (
-                    <Badge key={skill.id} variant="outline" className="flex items-center gap-1">
-                      {skill.name}
-                      <button onClick={() => removeSkill(skill.id)} className="ml-1 hover:bg-gray-300 rounded-full p-0.5">
-                        <X className="w-3 h-3" />
-                      </button>
-                    </Badge>
-                  ))}
-                </div>
-                
-                <div className="flex gap-2">
-                  <Select 
-                  // 1. open 상태를 직접 관리
-                  open={isSelectOpen}
-                  // 2. onOpenChange를 사용하여 상태 변경을 감지
-                  onOpenChange={setIsSelectOpen}
+              {/* 관심 분야 & 기술 스택 섹션 */}
+              <div className="space-y-6 p-6 border rounded-lg">
+                <h3 className="text-lg font-semibold flex items-center"><Heart className="w-5 h-5 mr-2" />관심 분야 & 기술 스택</h3>
+                {/* 관심 분야 */}
+                {/* <div className="space-y-4">
+                  <h4 className="font-medium text-gray-900">관심 분야</h4>
+                  <div className="flex flex-wrap gap-2">
+                    {profile.interests.map((interest) => (
+                      <Badge key={interest} variant="secondary" className="flex items-center gap-1">
+                        {interest}
+                        <button
+                          onClick={() => removeInterest(interest)}
+                          className="ml-1 hover:bg-gray-300 rounded-full p-0.5"
+                        >
+                          <X className="w-3 h-3" />
+                        </button>
+                      </Badge>
+                    ))}
+                  </div>
+                  <div className="flex gap-2">
+                    <Select value={newInterest} onValueChange={setNewInterest}>
+                      <SelectTrigger className="flex-1">
+                        <SelectValue placeholder="관심 분야 선택" />
+                      </SelectTrigger>
+                      <SelectContent>
+                        {availableInterests
+                          .filter((interest) => !profile.interests.includes(interest))
+                          .map((interest) => (
+                            <SelectItem key={interest} value={interest}>
+                              {interest}
+                            </SelectItem>
+                          ))}
+                      </SelectContent>
+                    </Select>
+                    <Button onClick={() => addInterest(newInterest)} disabled={!newInterest} size="sm">
+                      <Plus className="w-4 h-4" />
+                    </Button>
+                  </div>
+                  <p className="text-xs text-gray-500">관심 있는 공모전 분야를 선택해주세요</p>
+                </div> */}
 
-                  onValueChange={(skillName: string) => {
-                      const foundSkill = arraySkills.find((s) => s.name === skillName);
-                      if (foundSkill) {
-                        addSkill(foundSkill); // 찾은 Skills 객체를 addSkill에 전달
+                {/* 기술 스택 */}
+                <div className="space-y-4">
+                  <h4 className="font-medium text-gray-900">기술 스택</h4>
+                  <div className="flex flex-wrap gap-2">
+                    {selectSkill?.map((skill) => (
+                      <Badge key={skill.id} variant="outline" className="flex items-center gap-1">
+                        {skill.name}
+                        <button
+                          onClick={() => removeSkill(skill.id)}
+                          className="ml-1 hover:bg-gray-300 rounded-full p-0.5"
+                        >
+                          <X className="w-3 h-3" />
+                        </button>
+                      </Badge>
+                    ))}
+                  </div>
+                  <div className="flex gap-2">
+                    <Select 
+                      // 1. open 상태를 직접 관리
+                      open={isSelectOpen}
+                      // 2. onOpenChange를 사용하여 상태 변경을 감지
+                      onOpenChange={setIsSelectOpen}
 
-                        // --- 이 부분이 핵심입니다 ---
-                        // 아이템을 선택한 후에도 Select가 열려있도록 강제합니다.
-                        // onValueChange가 발생한 후 Select 내부적으로 닫히려고 할 수 있으므로,
-                        // 명시적으로 다시 열린 상태로 설정해 줍니다.
-                        setIsSelectOpen(true);
-                        
-                      }
-                      //console.log(isSelectOpen)
-                    }}>
+                      onValueChange={(skillName: string) => {
+                        const foundSkill = arraySkills.find((s) => s.name === skillName);
+                        if (foundSkill) {
+                          addSkill(foundSkill); // 찾은 Skills 객체를 addSkill에 전달
+
+                          // --- 이 부분이 핵심입니다 ---
+                          // 아이템을 선택한 후에도 Select가 열려있도록 강제합니다.
+                          // onValueChange가 발생한 후 Select 내부적으로 닫히려고 할 수 있으므로,
+                          // 명시적으로 다시 열린 상태로 설정해 줍니다.
+                          setIsSelectOpen(true);
+                          
+                        }
+                        //console.log(isSelectOpen)
+                      }}>
                       
                     <SelectTrigger className="flex-1">
                       <SelectValue placeholder="기술 스택 선택" />
@@ -417,24 +391,24 @@ function ProfileEditContent() {
                         ))}
                     </SelectContent>
                   </Select>
+                    {/* <Button onClick={() => addSkill(newSkill)} disabled={!newSkill} size="sm">
+                      <Plus className="w-4 h-4" />
+                    </Button> */}
+                  </div>
+                  <p className="text-xs text-gray-500">보유하고 있는 기술이나 도구를 선택해주세요</p>
                 </div>
-              </CardContent>
-            </Card>
+              </div>
 
-            {/* 추가 정보 */}
-            <Card>
-              <CardHeader>
-                <CardTitle>추가 정보</CardTitle>
-              </CardHeader>
-              <CardContent className="space-y-4">
-                <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+              {/* 추가 정보 섹션 */}
+              <div className="space-y-6 p-6 border rounded-lg">
+                <h3 className="text-lg font-semibold flex items-center"><Code className="w-5 h-5 mr-2" />추가 정보 (선택)</h3>
+                <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
                   <div className="space-y-2">
                     <Label htmlFor="education">학력</Label>
                     <Input
                       id="education"
-                      value={profile?.education}
-                      onChange={(e) =>setProfile(prev => ({ ...prev, education: e.target.value }))}
-                      
+                      value={profile.education}
+                      onChange={(e) => setProfile({ ...profile, education: e.target.value })}
                       placeholder="예: 컴퓨터공학과 학사"
                     />
                   </div>
@@ -442,20 +416,20 @@ function ProfileEditContent() {
                     <Label htmlFor="experience">경력</Label>
                     <Input
                       id="experience"
-                      value={profile?.experience}
-                      onChange={(e) => setProfile(prev => ({ ...prev, experience: e.target.value }))}
+                      value={profile.experience}
+                      onChange={(e) => setProfile({ ...profile, experience: e.target.value })}
                       placeholder="예: 프론트엔드 개발자 2년"
                     />
                   </div>
-                  <div className="space-y-2">
+                  {/* <div className="space-y-2">
                     <Label htmlFor="portfolio">포트폴리오 URL</Label>
                     <Input
                       id="portfolio"
-                      value={profile?.portfolioUrl}
-                      onChange={(e) => setProfile(prev => ({ ...prev, portfolioUrl: e.target.value }))}
+                      value={profile.portfolio}
+                      onChange={(e) => setProfile({ ...profile, portfolio: e.target.value })}
                       placeholder="https://portfolio.example.com"
                     />
-                  </div>
+                  </div> */}
                   {/* <div className="space-y-2">
                     <Label htmlFor="github">GitHub URL</Label>
                     <Input
@@ -466,33 +440,45 @@ function ProfileEditContent() {
                     />
                   </div> */}
                 </div>
-              </CardContent>
-            </Card>
-          </div>
-        </div>
-        {/* 저장 버튼 */}
-        <div className="mt-8 flex justify-end">
-        <Button onClick={handleSave} disabled={isLoading}>
-            {isLoading ? (
-              <>
-                <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-white mr-2"></div>
-                저장 중...
-              </>
-            ) : (
-              <>
-                <Save className="w-4 h-4 mr-2" />
-                저장하기
-              </>
-            )}
-          </Button>
+              </div>
+
+              <Alert>
+                <Sparkles className="h-4 w-4" />
+                <AlertDescription>
+                  프로필을 더 자세히 작성할수록 더 정확한 공모전 추천을 받을 수 있습니다!
+                </AlertDescription>
+              </Alert>
+
+              {/* 최종 액션 버튼 */}
+              <div className="flex flex-col sm:flex-row gap-3 pt-6">
+                <Button variant="outline" onClick={handleSkip} className="flex-1 bg-transparent">
+                  <Skip className="w-4 h-4 mr-2" />
+                  나중에 설정
+                </Button>
+                <Button onClick={handleSave} disabled={isLoading} className="flex-1">
+                  {isLoading ? (
+                    <>
+                      <Loader2 className="w-4 h-4 mr-2 animate-spin" />
+                      저장 중...
+                    </>
+                  ) : (
+                    <>
+                      <Save className="w-4 h-4 mr-2" />
+                      프로필 저장
+                    </>
+                  )}
+                </Button>
+              </div>
+            </CardContent>
+          </Card>
         </div>
       </div>
 
-      {/* 푸터 */}
       <Footer />
     </div>
   )
 }
+
 
 
 export default function ProfileEditPage() {
